src/requirements.txt
python-telegram-bot[http2,rate-limiter,job-queue]==22.0
pyyaml==6.0.2
